move initializeStorage and pollUsers to functions
